refactor(tui): extract V1 theme definitions
